@@ -1,6 +1,9 @@
 package com.example.myapplication
 
+import CalendarScrn
+import HomeScrn
 import SplashScreen
+import TasksViewModel
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
@@ -15,11 +18,12 @@ import com.example.myapplication.ui.screen.Otp.OptScrn
 import com.example.myapplication.ui.screen.RecoverPassword.RecoverPasswordScrn
 import com.example.myapplication.ui.screen.SignIn.SignInScrn
 import com.example.myapplication.ui.screen.SignUp.SignUpScrn
-
 import com.example.myapplication.ui.theme.MatuleTheme
 import kotlinx.serialization.Serializable
 import org.koin.android.ext.android.get
-
+import org.koin.androidx.viewmodel.ext.android.viewModel
+import com.example.myapplication.ui.data.remote.Tasks.TaskRepository
+import com.example.myapplication.ui.data.remote.Tasks.TasksViewModelFactory
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -29,8 +33,12 @@ class MainActivity : ComponentActivity() {
             val navController = rememberNavController()
             val dataStore = DataStoreOnBoarding(LocalContext.current)
             val authUseCase: AuthUseCase = get()
+
+            // Получаем ViewModel через Koin
+            val tasksViewModel: TasksViewModel by viewModel()
+
             MatuleTheme {
-                NavHost(navController, startDestination = SplashScreen){
+                NavHost(navController, startDestination = SplashScreen) {
 
                     composable<SplashScreen> {
                         SplashScreen(
@@ -54,15 +62,14 @@ class MainActivity : ComponentActivity() {
                             navController = navController,
                             onSignInSuccess = {
                                 navController.navigate(route = Otp)
-
                             }
                         )
                     }
-//
+
                     composable<RecoverPassword> {
                         RecoverPasswordScrn(navController)
                     }
-//
+
                     composable<Registration> {
                         SignUpScrn(
                             onNavigationToHome = {
@@ -73,9 +80,26 @@ class MainActivity : ComponentActivity() {
                     }
 
                     composable<Otp> {
-                         OptScrn {
+                        OptScrn {
                             navController.navigate(route = Home)
                         }
+                    }
+
+                    composable<Home> {
+                        val tasksViewModel: TasksViewModel by viewModel()
+                        HomeScrn(
+                            onNavigateToCalendar = {
+                                navController.navigate(route = Calendar)
+                            },
+                            tasksViewModel = tasksViewModel
+                        )
+                    }
+
+                    composable<Calendar> {
+                        CalendarScrn(
+                            onBack = { navController.popBackStack() },
+                            tasksViewModel = tasksViewModel
+                        )
                     }
                 }
             }
@@ -99,3 +123,5 @@ object Otp
 object Slides
 @Serializable
 object Home
+@Serializable
+object Calendar
