@@ -1,8 +1,8 @@
 package com.example.myapplication.ui.data
 
-import com.example.myapplication.ui.data.remote.Auth
+import com.example.myapplication.ui.data.remote.dto.response.Auth
 import com.example.myapplication.ui.data.remote.User
-import com.example.myapplication.ui.data.remote.dto.LoginRequest
+import com.example.myapplication.ui.data.remote.Request.LoginRequest
 import com.example.myapplication.ui.data.remote.dto.response.TokenResponse
 import kotlinx.coroutines.delay
 
