@@ -17,6 +17,7 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.systemBarsPadding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.BasicTextField
@@ -81,7 +82,7 @@ fun OptScrn(title: String = "", navigation: () -> Unit) {
     }
 
     Scaffold(
-        modifier = Modifier.fillMaxSize().background(MatuleTheme.colors.biskuit),
+        modifier = Modifier.fillMaxSize().background(MatuleTheme.colors.biskuit).systemBarsPadding(),
         topBar = {
             Row(
                 modifier = Modifier
