@@ -12,6 +12,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.systemBarsPadding
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Scaffold
@@ -43,7 +44,7 @@ fun RecoverPasswordScrn(navController: NavController) {
     val recoverPasswordViewModel: RecoverPasswordViewModel = viewModel()
     var isDialogOpen by remember { mutableStateOf(false) }
 
-    Scaffold(
+    Scaffold(modifier = Modifier.systemBarsPadding(),
         topBar = {
             Row(
                 modifier = Modifier
